Add support for another generic 2.4" 320x240 DPF
This is exactly dpf-ax-acme_8.patch from
http://sourceforge.net/p/dpf-ax/patches/5/
but using acme_10 instead of acme_8 directory name.

This photo frame is similar to acme_4 but with different LCD
initialization. The original firmware had no backlight control,
turning it on 100% via the port and not using the timer. I set up
a backlight table based on the sRGB curve and enabled backlight
control. There is no contrast control in the original firmware,
and its contrast table is useless. I don't know how to add
contrast control. The LCD controller is similar to Ilitek ILI9431.

Since there is no brand or model shown anywhere, I didn't know
how to name this and just used the next free acme designation.